ex4.15-4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,29 +1,63 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blogs')
+const User = require('../models/users')
+const jwt = require('jsonwebtoken')
 
-blogsRouter.get('/', (request, response) => {
-	Blog
-		.find({})
-		.then(blogs => {
-			response.json(blogs)
-		})
+blogsRouter.get('/', async (request, response) => {
+	const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
+	response.json(blogs)
 })
 
-blogsRouter.post('/', (request, response) => {
-	const blog = new Blog(request.body)
+blogsRouter.get('/:id', async (request, response) => {
+	const blog = await Blog.findById(request.params.id)
+	response.json(blog)
+})
 
-	blog
-		.save()
-		.then(result => {
-			response.status(201).json(result)
-		}).catch(error => {
-			response.status(400).json(error)
+blogsRouter.post('/', async (request, response) => {
+	const body = request.body
+
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	// console.log("DEBUG ::: ", decodedToken)
+	if (!decodedToken.id) {
+		return response.status(401).json({
+			error: 'token invalid'
 		})
+	}
+	const user = await User.findById(decodedToken.id)
+
+	const blog = new Blog({
+		title: body.title,
+		author: body.author,
+		url: body.url,
+		likes: body.likes,
+		user: user._id
+	})
+
+	// try catch error handling block is eliminated due to the express-async-error module
+	const savedBlog = await blog.save()
+	user.blogs = user.blogs.concat(savedBlog._id)
+	await user.save()
+
+	response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-	await Blog.findByIdAndRemove(request.params.id)
-	response.status(204).end()
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	if (!decodedToken.id) {
+		return response.status(401).json({
+			error: 'token invalid'
+		})
+	}
+	const user = await User.findById(decodedToken.id)
+
+	const blog = await Blog.findById(request.params.id)
+
+	if (blog.user.toString() === user._id.toString()) {
+		await Blog.findByIdAndRemove(request.params.id)
+		response.status(204).end()
+	} else {
+		response.status(401).json({ 'error': 'unauthorized' })
+	}
 })
 
 blogsRouter.put('/:id', async (request, response) => {
